add function deleted to tasks component

+++ src/apps/shared/components/delet/delet.component.ts
@@ -9,34 +9,40 @@ import { ToastrService } from 'ngx-toastr';
   styleUrls: ['./delet.component.scss'],
 })
 export class DeletComponent {
+  source: string = '';
 
   constructor(
     public dialogRef: MatDialogRef<DeletComponent>,
     @Inject(MAT_DIALOG_DATA) public data: any,
     private _ActivatedRoute: ActivatedRoute,
-    private _SharedService:SharedService,
-    private _ToastrService:ToastrService
+    private _SharedService: SharedService,
+    private _ToastrService: ToastrService
   ) {
-  console.log(data);
-  
+    console.log('dialog: ', data);
+    this.source = data.source;
+    console.log('source: ', this.source);
   }
-  
+
   onNoClick(): void {
     this.dialogRef.close();
   }
 
-  delete(id:number){
-this._SharedService.deleteprojects(id).subscribe({
-  next:(res)=>{
-console.log(res);
-    
-this.dialogRef.close(true);
-  },error:()=>{
+  // source = projects || tasks
+  delete(id: number) {
+    if (this.source === 'projects') {
+      this._SharedService.deleteprojects(id).subscribe({
+        next: (res) => {
+          console.log(res);
 
-  },complete:()=>{
-      this._ToastrService.success('Deleted successfully');
+          this.dialogRef.close(true);
+        },
+        error: () => {},
+        complete: () => {
+          this._ToastrService.success('Deleted successfully');
+        },
+      });
+    } else {
+      // هتحذف tasks
+    }
   }
-})
-  }
-
 }
